refactor(InstallControls): move the main component to a named module
